+++ src/briefy/choreographer/actions/mail/leica/comment.py
@@ -6,6 +6,8 @@ from briefy.choreographer.events.leica import comment as events
 from zope.component import adapter
 from zope.interface import implementer
 
+import typing as t
+
 
 class CommentMail(LeicaMail):
     """Base class for emails sent on Comment events."""
@@ -14,41 +16,15 @@ class CommentMail(LeicaMail):
     """Name of the entity to be processed here."""
 
     @property
-    def action_url(self):
+    def action_url(self) -> str:
         """Action URL."""
         return self._action_url
 
-    def _recipients(self, field_name: str):
-        """Return a list of valid recipients."""
-        data = self.data['entity']
-        recipients = []
-        if field_name == 'last_transition':
-            history = data['state_history']
-            actor = history[0]['actor']
-            users = [actor, ]
-        else:
-            users = data[field_name]
-        for user in users:
-            if not user['internal']:
-                continue
-            recipients.append(
-                {
-                    'first_name': user['first_name'],
-                    'fullname': user['fullname'],
-                    'email': user['email'],
-                }
-            )
-        return recipients
-
-    def transform(self) -> list:
+    def transform(self) -> t.List[dict]:
         """Transform data."""
-        base_payload = super().transform()
+        base_payload = super().transform()[0]
         data = self.data
-        recipients = self.recipient
-        if isinstance(recipients, dict):
-            recipients = [recipients, ]
-        elif not recipients:
-            return []
+        recipients = self.recipients
         payload = []
         for recipient in recipients:
             payload_item = {}
@@ -78,7 +54,7 @@ class CommentCustomerMail(CommentMail):
     """Base class for emails sent to the Customer on Comment events."""
 
     @property
-    def recipient(self):
+    def recipients(self) -> t.List[dict]:
         """Return the data to be used as the recipient of this message."""
         return self._recipients('customer_users')
 
@@ -87,7 +63,7 @@ class CommentCreativeMail(CommentMail):
     """Base class for emails sent to the Creative on Comment events."""
 
     @property
-    def recipient(self):
+    def recipients(self) -> t.List[dict]:
         """Return the data to be used as the recipient of this message."""
         return self._recipients('professional_user')
 
@@ -96,7 +72,7 @@ class CommentPMMail(CommentMail):
     """Base class for emails sent to the PMs on Comment events."""
 
     @property
-    def recipient(self):
+    def recipients(self) -> t.List[dict]:
         """Return the data to be used as the recipient of this message."""
         return self._recipients('project_managers')
 
@@ -111,13 +87,11 @@ class CommentCreatedToCreative(CommentCreativeMail):
     subject = '{COMMENTER_FIRSTNAME} has commented on your assignment'
 
     @property
-    def action_url(self):
+    def action_url(self) -> str:
         """Action URL."""
         data = self.data
-        return '{base}assignments/{id}'.format(
-            id=data['entity']['id'],
-            base=PLATFORM_URL
-        )
+        id_ = data['entity']['id']
+        return f'{PLATFORM_URL}assignments/{id_}'
 
     @property
     def available(self) -> bool:
@@ -138,13 +112,11 @@ class CommentCreatedToCustomer(CommentCustomerMail):
     subject = '{COMMENTER_FIRSTNAME} has commented on your order'
 
     @property
-    def action_url(self):
+    def action_url(self) -> str:
         """Action URL."""
         data = self.data
-        return '{base}orders/{id}'.format(
-            id=data['entity']['id'],
-            base=PLATFORM_URL
-        )
+        id_ = data['entity']['id']
+        return f'{PLATFORM_URL}orders/{id_}'
 
     @property
     def available(self) -> bool:
@@ -165,13 +137,11 @@ class CommentCreatedByCustomerToPM(CommentPMMail):
     subject = '{COMMENTER_FIRSTNAME} commented on an order'
 
     @property
-    def action_url(self):
+    def action_url(self) -> str:
         """Action URL."""
         data = self.data
-        return '{base}orders/{id}'.format(
-            id=data['entity']['id'],
-            base=PLATFORM_URL
-        )
+        id_ = data['entity']['id']
+        return f'{PLATFORM_URL}orders/{id_}'
 
     @property
     def available(self) -> bool: