+++ src/briefy/choreographer/actions/mail/quote.py
@@ -7,6 +7,8 @@ from briefy.choreographer.events import lead
 from zope.component import adapter
 from zope.interface import implementer
 
+import typing as t
+
 
 class QuoteMail(Mail):
     """Base class for emails sent on Quote events."""
@@ -20,29 +22,23 @@ class QuoteMail(Mail):
 
         :returns: Dictionary with two keys - name, email
         """
-        return {
-            'name': MAIL_ACTION_QUOTE_SENDER_NAME,
-            'email': MAIL_ACTION_QUOTE_SENDER_EMAIL,
-        }
+        return {'name': MAIL_ACTION_QUOTE_SENDER_NAME, 'email': MAIL_ACTION_QUOTE_SENDER_EMAIL}
 
-    def transform(self) -> dict:
+    def transform(self) -> t.List[dict]:
         """Transform data."""
         payload = super().transform()
         data = self.data
-        fullname = '{first} {last}'.format(
-            first=data['first_name'],
-            last=data['last_name']
-        )
-        payload['fullname'] = fullname
-        payload['email'] = data['email']
-        payload['company'] = data['company']
-        payload['phone_number'] = data['phone_number']
-        payload['company_site'] = data.get('company_site', '')
+        fist_name = data['first_name']
+        last_name = data['last_name']
+        fullname = f'{fist_name} {last_name}'
+        payload[0]['fullname'] = fullname
+        payload[0]['email'] = data['email']
+        payload[0]['company'] = data['company']
+        payload[0]['phone_number'] = data['phone_number']
+        payload[0]['company_site'] = data.get('company_site', '')
 
-        payload['data'] = {
-            'FULLNAME': fullname,
-            'EMAIL': data['email'],
-            'SUBJECT': self.subject,
+        payload[0]['data'] = {
+            'FULLNAME': fullname, 'EMAIL': data['email'], 'SUBJECT': self.subject
         }
         return payload