+++ src/briefy/choreographer/actions/slack/leica/assignment.py
@@ -2,8 +2,11 @@
 from briefy.choreographer.actions.slack import ISlack
 from briefy.choreographer.actions.slack import Slack
 from briefy.choreographer.events.leica import assignment as events
+from briefy.choreographer.utils.user_data import users_data_by_role
 from zope.component import adapter
 from zope.interface import implementer
+
+import typing as t
 
 
 class AssignmentSlack(Slack):
@@ -15,18 +18,15 @@ class AssignmentSlack(Slack):
     title = 'Assignment'
     text = 'New Assignment!!'
 
-    def transform(self) -> dict:
+    def transform(self) -> t.List[dict]:
         """Transform data."""
         payload = super().transform()
+        payload_item = payload[0]
         data = self.data
-        payload['title'] = self.title
-        payload['text'] = (
-            'Assignment can be seen <{url}|here>'.format(
-                url=self._action_url,
-            )
-        )
-        payload['username'] = 'Briefy Bot'
-        payload['data'] = {
+        payload_item['title'] = self.title
+        payload_item['text'] = f'Assignment can be seen <{self._action_url}|here>'
+        payload_item['username'] = 'Briefy Bot'
+        payload_item['data'] = {
             'fields': [
                 {'title': 'Project',
                  'value': data.get('project', {}).get('title'),
@@ -36,7 +36,7 @@ class AssignmentSlack(Slack):
         }
         professional = data.get('professional')
         if professional:
-            payload['data']['fields'].append(
+            payload_item['data']['fields'].append(
                 {'title': 'Creative',
                  'value': data.get('professional', {}).get('title'),
                  'short': True,
@@ -252,22 +252,22 @@ class AssignmentWfAssignQAManager(AssignmentSlack):
 
     title = 'QA Manager started the review'
 
-    def transform(self) -> dict:
+    def transform(self) -> t.List[dict]:
         """Transform data."""
         payload = super().transform()
-        data = self.data
-        qa_managers = data['qa_managers']
-        if qa_managers:
-            qa_manager = qa_managers[0]
-            fields = payload['data']['fields']
+        payload_item = payload[0]
+        internal_qa = users_data_by_role(self.event, 'assignment_internal_qa')
+        if internal_qa:
+            internal_qa = internal_qa[0]
+            fields = payload_item['data']['fields']
             fields.append(
                 {
                     'title': 'QA Manager',
-                    'value': qa_manager.get('fullname', ''),
+                    'value': internal_qa.get('fullname', ''),
                     'short': True,
                 },
             )
-            payload['data']['fields'] = fields
+            payload_item['data']['fields'] = fields
         return payload