+++ tests/action/email/test_email_quote.py
@@ -1,7 +1,7 @@
 """Briefy email action for Quote events tests."""
 from briefy.choreographer.actions.mail import IMail
 from briefy.choreographer.actions.mail.quote import QuoteCreated
-from briefy.choreographer.events.lead import lead as events
+from briefy.choreographer.events import lead as events
 from conftest import BaseActionCase
 
 
@@ -18,6 +18,9 @@ class TestQuoteCreated(BaseActionCase):
         """Test data transform."""
         obj = self.obj
         payload = obj.transform()
+        assert isinstance(payload, list)
+        assert len(payload) == 1
+        payload = payload[0]
         data = payload['data']
         assert isinstance(payload['sender_name'], str)
         assert isinstance(payload['sender_email'], str)